Created helper functions
