+++ local_test.js
@@ -2,6 +2,6 @@ import clipper from "./index.js";
 import clipperOld from "./index copy.js";
 
 // clipper("www.whatever.com", "shorten");
-clipper("www.google.com/maps/place/Big+Ben/", "shorten", "shorten");
+clipper("www.google.com/maps/place/Big+Ben/", "remove", "shorten");
 // clipper(34654775676, "none", "shorten");
 // clipperOld("www.whatever.com", "none", "shorten");
