+++ plan.md
@@ -1,6 +1,6 @@
 ## Version 2.0.0 to do:
 
-1. Replace console logs with errors
+1. ✅ Replace console logs with errors, and add in try-catch blocks
    -adjust tests accordingly
 2. Pull the little functions out into separate components (helpers)
 3. Input validation to check whether the URL is a string.
